Tidy & Format (#112)

+++ src/obfuscator.cpp
@@ -4,15 +4,14 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2021 Datadog, Inc.
 
+#include <cstddef>
+#include <log.hpp>
 #include <obfuscator.hpp>
 #include <utils.h>
-#include <log.hpp>
 
-namespace ddwaf
-{
+namespace ddwaf {
 
-namespace
-{
+namespace {
 
 bool match(re2::RE2 &regex, std::string_view value)
 {
@@ -20,13 +19,13 @@ bool match(re2::RE2 &regex, std::string_view value)
     return regex.Match(key_ref, 0, value.size(), re2::RE2::UNANCHORED, nullptr, 0);
 }
 
-}
+} // namespace
 
-obfuscator::obfuscator(std::string_view key_regex_str,
-    std::string_view value_regex_str)
+obfuscator::obfuscator(std::string_view key_regex_str, std::string_view value_regex_str)
 {
     re2::RE2::Options options;
-    options.set_max_mem(512 * 1024);
+    // NOLINTNEXTLINE(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
+    options.set_max_mem(static_cast<int64_t>(512 * 1024));
     options.set_log_errors(false);
     options.set_case_sensitive(false);
 
@@ -34,20 +33,16 @@ obfuscator::obfuscator(std::string_view key_regex_str,
         re2::StringPiece sp(key_regex_str.data(), key_regex_str.size());
         key_regex = std::make_unique<re2::RE2>(sp, options);
 
-        if (!key_regex->ok())
-        {
-            DDWAF_ERROR("invalid obfuscator key regex: %s - using default",
-                key_regex->error_arg().c_str());
+        if (!key_regex->ok()) {
+            DDWAF_ERROR(
+                "invalid obfuscator key regex: %s - using default", key_regex->error_arg().c_str());
 
-            sp = re2::StringPiece(default_key_regex_str.data(),
-                default_key_regex_str.size());
+            sp = re2::StringPiece(default_key_regex_str.data(), default_key_regex_str.size());
             key_regex = std::make_unique<re2::RE2>(sp, options);
 
-            if (!key_regex->ok())
-            {
+            if (!key_regex->ok()) {
                 throw std::runtime_error(
-                    "invalid default obfuscator key regex: " +
-                    key_regex->error_arg());
+                    "invalid default obfuscator key regex: " + key_regex->error_arg());
             }
         }
     }
@@ -56,10 +51,8 @@ obfuscator::obfuscator(std::string_view key_regex_str,
         const re2::StringPiece sp(value_regex_str.data(), value_regex_str.size());
         value_regex = std::make_unique<re2::RE2>(sp, options);
 
-        if (!value_regex->ok())
-        {
-            DDWAF_ERROR("invalid obfuscator value regex: %s",
-                value_regex->error_arg().c_str());
+        if (!value_regex->ok()) {
+            DDWAF_ERROR("invalid obfuscator value regex: %s", value_regex->error_arg().c_str());
         }
     }
 }
@@ -69,10 +62,9 @@ bool obfuscator::is_sensitive_key(std::string_view key) const
     return key_regex ? match(*key_regex, key) : false;
 }
 
-
 bool obfuscator::is_sensitive_value(std::string_view value) const
 {
     return value_regex ? match(*value_regex, value) : false;
 }
 
-}
+} // namespace ddwaf
